[Silver II] Title: 이상한 술집, Time: 152 ms, Memory: 17308 KB -BaekjoonHub

+++ 백준/Silver/13702. 이상한 술집/이상한 술집.java
@@ -1,5 +1,4 @@
 
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -11,16 +10,16 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st=new StringTokenizer(br.readLine());
 
+        //int->long : 막걸리의 용량은 pow(2,31)
+        //주전자의 용량, 최대막걸리 용량이 pow(2,31)인 경우 low=hight+1
         int n = Integer.parseInt(st.nextToken());
         int m = Integer.parseInt(st.nextToken());
-        long[] arr = new long[n];
-        long max = 0;
+        int[] arr = new int[n];
+        int max = 0;
         for (int i = 0; i < n; i++) {
             arr[i] = Integer.parseInt(br.readLine());
             max = Math.max(max, arr[i]);
         }
-
-        //n과 m이 0일 경우
 
         long low = 1, high = max;
         while (low <= high) {
